+++ CHANGELOG.md
@@ -1,7 +1,10 @@
-## [v0.8.2.0](https://github.com/LambdaHack/LambdaHack/compare/v0.8.1.2...v0.8.2.0)
+## [v0.8.3.0](https://github.com/LambdaHack/LambdaHack/compare/v0.8.1.2...v0.8.3.0)
 
-- Add a hack to run SDL2 on OS X in the main thread, fixing the crash
+- Add a hack to run SDL2 on the main thread, fixing the OS X crash
+- Warn visually when impressed and Calm running low, risking domination
+- Display actor as red when low Calm and impressed or when low HP
 - Fix, complete and fine tune UI, AI and server skill and weapon checks
+- Fix a bug where item aspects look different to clients than to the server
 - Change the requirements for the main menu ASCII art
 
 ## [v0.8.1.2](https://github.com/LambdaHack/LambdaHack/compare/v0.8.1.1...v0.8.1.2)